Add local notifications for bad posture
Posture notifications will only be sent when the app is in the background. In addition, after one notification is sent, no additional posture notifications will be sent until the next time the app goes to the background. This will not annoy the user with multiple notifications because the hardware device is constantly streaming.