Keep posture activity running after tapping the notification
Before, tapping the notification from the posture foreground service would disable
the posture activity module because `SensorDataService` issues the
`STOP_POSTURE_FOREGROUND_SERVICE` intent in `onActivityResumed`, and the foreground
service had logic to kill the posture activity when receiving the
`STOP_POSTURE_FOREGROUND_SERVICE` intent. This change updates the logic for
`STOP_POSTURE_FOREGROUND_SERVICE` so that it will only stop the foreground service
and not kill the posture activity.

A new `STOP_POSTURE_ACTIVITY` is added and would be used to explicitly kill the
posture activity.