Replace ternary operators, relocate initialization of default values
The ternary operators do not provide flexibility in the future when we add additional
metrics aside from height and weight.

In addition, default values and units will be set in the ProfilePicker component
instead of the OnBoarding component's state since having it in the state would
automatically initialize the values without the user selecting the values.