Do not unregister broadcast receiver on activity pause
The broadcast receiver needs to emit a JS event when the Bluetooth adapter state changes,
and if the Android prompt comes up to ask the user to enable Bluetooth, the activity is
paused. As a result, the broadcast receiver must still be registered to the context even
while the activity is paused.