Use separate flag for tracking a connect attempt
Currently, the `connect` and `getInfo` actions both modify the `inProgress`
flag in the device reducer slice. There are situations where `inProgress`
is set to `false` from `getInfo` before the device is even connected. As
a result, on app load or reconnecting when the app is back in the foreground
with a saved device, the banner will not be able to show it's in the process
of connecting to the device because a `getInfo` action is dispatched before
the banner is rendered and before the `connect` action finishes. So when the
banner is rendered, the `inProgress` flag would be `false` but the `connect`
action still hasn't dispatched yet.