Convert TitleBar to a container component
When it was a presentational component, requiring route/title bar data to be passed down from the `Application` container component, it would cause navigation to be sluggish. This was due to the `Application` component having to re-render more than necessary because it was listening to changes in the entire `app` reducer slice.

Now that `TitleBar` is a container component, it can selectively listen to only the `titleBar` object from the `app` reducer slice, and `Application` doesn't have to listen to that slice.